requirements.txt
# Core deep learning stack
torch>=2.2.0,<2.4.0
torchvision>=0.17.0,<0.19.0

# Configuration management
pyyaml>=6.0.0,<7.0.0
matplotlib>=3.8.0,<4.0.0
